requirements.txt
flask
yt-dlp
openai-whisper
torch --index-url https://download.pytorch.org/whl/cpu
setuptools